requirements.txt
click==8.1.3
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap4==4.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
visitor==0.1.3
Werkzeug==2.1.2
